requirements.txt
torch~=2.7.1
torchvision~=0.22.1
Pillow~=11.2.1
matplotlib~=3.10.3
#streamlit~=1.48.0
numpy~=2.2.6
pytest
ruff
black
mypy